requirements.txt
spotipy==2.23.0
python-dotenv==1.0.0
urllib3==1.26.7